Move search `Selector` to separate file
