+++ src/containers/SearchPage/components/form/SearchImageForm.tsx
@@ -15,8 +15,9 @@ import { getLicensesWithTranslations } from '../../../../util/licenseHelpers';
 import { SearchParams } from './SearchForm';
 import { SubjectType } from '../../../../modules/taxonomy/taxonomyApiInterfaces';
 import { useLicenses } from '../../../../modules/draft/draftQueries';
-import GenericSearchForm, { OnFieldChangeFunction, SearchFormSelector } from './GenericSearchForm';
+import GenericSearchForm, { OnFieldChangeFunction } from './GenericSearchForm';
 import { TagType } from './SearchTagGroup';
+import { SearchFormSelector } from './Selector';
 
 interface Props {
   search: (o: SearchParams) => void;
