+++ src/containers/SearchPage/components/form/Selector.tsx
@@ -1,0 +1,105 @@
+/**
+ * Copyright (c) 2022-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+import { useTranslation } from 'react-i18next';
+import { unreachable } from '../../../../util/guards';
+import ObjectSelector from '../../../../components/ObjectSelector';
+import CheckboxSelector from './CheckboxSelector';
+import InlineDatePicker from '../../../FormikForm/components/InlineDatePicker';
+import { SearchParams } from './SearchForm';
+import { OnFieldChangeFunction } from './GenericSearchForm';
+
+interface SearchFormSelectorBase {
+  parameterName: keyof SearchParams;
+  value?: string;
+  width?: number;
+}
+
+/** These types are to extract keys of a specific type from SearchParams */
+type RSP = Required<SearchParams>;
+type SearchParamsWithOnlyType<P> = { [k in keyof RSP]: RSP[k] extends P ? k : never }[keyof RSP];
+type SearchParamsOfType<O> = { [k in SearchParamsWithOnlyType<O>]: O };
+
+export type SearchFormSelector =
+  | CheckboxSelectorType
+  | ObjectSelectorType
+  | DatePickerSelectorType
+  | TextInputSelectorType;
+
+interface CheckboxSelectorType extends SearchFormSelectorBase {
+  formElementType: 'check-box';
+  parameterName: keyof SearchParamsOfType<boolean>;
+}
+
+interface ObjectSelectorType extends SearchFormSelectorBase {
+  formElementType: 'dropdown';
+  options: { id: string; name: string }[];
+  parameterName: keyof SearchParamsOfType<string>;
+}
+
+interface DatePickerSelectorType extends SearchFormSelectorBase {
+  formElementType: 'date-picker';
+  parameterName: keyof SearchParamsOfType<string>;
+}
+
+interface TextInputSelectorType extends SearchFormSelectorBase {
+  formElementType: 'text-input';
+}
+
+interface SelectorProps {
+  selector: SearchFormSelector;
+  onFieldChange: OnFieldChangeFunction;
+  searchObject: SearchParams;
+}
+
+const Selector = ({ selector, onFieldChange, searchObject }: SelectorProps) => {
+  const { t } = useTranslation();
+  switch (selector.formElementType) {
+    case 'text-input':
+      // TODO: text-input/query is handled specifically in `GenericSearchForm`
+      //       in the future this should probably be moved here.
+      //       for now this branch will remain here to satisfy the typing.
+      return null;
+    case 'date-picker':
+      const datePickerValue = searchObject[selector.parameterName];
+      return (
+        <InlineDatePicker
+          name={selector.parameterName}
+          onChange={e => onFieldChange(selector.parameterName, e.target.value)}
+          placeholder={t(`searchForm.types.${selector.parameterName}`)}
+          value={datePickerValue ?? ''}
+        />
+      );
+    case 'check-box':
+      const checkboxValue = searchObject[selector.parameterName];
+      return (
+        <CheckboxSelector
+          name={selector.parameterName}
+          checked={checkboxValue ?? false}
+          onChange={e => onFieldChange(selector.parameterName, e.currentTarget.checked)}
+        />
+      );
+    case 'dropdown':
+      const dropdownValue = searchObject[selector.parameterName];
+      return (
+        <ObjectSelector
+          name={selector.parameterName}
+          value={dropdownValue ?? ''}
+          options={selector.options}
+          idKey="id"
+          labelKey="name"
+          emptyField
+          onChange={evt => onFieldChange(selector.parameterName, evt.currentTarget.value)}
+          placeholder={t(`searchForm.types.${selector.parameterName}`)}
+        />
+      );
+    default:
+      return unreachable(selector);
+  }
+};
+
+export default Selector;
